Update the 'args' module
