--- args/src/main/kotlin/com/hexagonkt/args/ArgsManager.kt
@@ -1,6 +1,0 @@
-package com.hexagonkt.args
-
-object ArgsManager {
-    var groupOptions: Boolean = false
-    var repeatParameter: Boolean = true
-}
