+++ args/src/main/kotlin/com/hexagonkt/args/Option.kt
@@ -3,11 +3,20 @@ package com.hexagonkt.args
 import kotlin.reflect.KClass
 
 data class Option<T : Any>(
-    val parameter: Parameter<T>,
-    val shortName: Char? = null,
-) : Property<T> by parameter {
+    override val type: KClass<T>,
+    override val names: LinkedHashSet<String>,
+    override val description: String? = null,
+    override val regex: Regex? = null,
+    override val optional: Boolean = true,
+    override val multiple: Boolean = false,
+    override val values: List<T> = emptyList(),
+) : Property<T> {
 
-    private val hasValue: Boolean = type != Boolean::class
+    val value: T? = values.firstOrNull()
+
+    companion object {
+        val optionRegex = "([A-Za-z0-9]|[a-z0-9\\-]{2,})".toRegex()
+    }
 
     constructor(
         type: KClass<T>,
@@ -18,7 +27,15 @@ data class Option<T : Any>(
         optional: Boolean = true,
         multiple: Boolean = false,
         values: List<T> = emptyList(),
-    ) : this(Parameter(type, name, description, regex, optional, multiple, values), shortName)
+    ) : this(
+        type,
+        LinkedHashSet(listOfNotNull(shortName?.toString(), name)),
+        description,
+        regex,
+        optional,
+        multiple,
+        values
+    )
 
     constructor(
         type: KClass<T>,
@@ -29,63 +46,11 @@ data class Option<T : Any>(
         optional: Boolean = true,
         multiple: Boolean = false,
         value: T,
-    ) : this(Parameter(type, name, description, regex, optional, multiple, listOf(value)), shortName)
+    ) : this(type, shortName, name, description, regex, optional, multiple, listOf(value))
 
     init {
-        require(shortName?.isLetterOrDigit() ?: true) {
-            "Short name must be a letter or a digit: $shortName"
-        }
-        require(shortName != null || name != null) {
-            "Option requires a short name, a long name or both. Short: $shortName, Long: $name"
+        require(names.all { it.matches(optionRegex) }) {
+            "Names must comply with ${optionRegex.pattern} regex: $names"
         }
     }
-
-    fun copy(
-        type: KClass<T> = this.type,
-        name: String? = this.name,
-        description: String? = this.description,
-        regex: Regex? = this.regex,
-        optional: Boolean = this.optional,
-        multiple: Boolean = this.multiple,
-        values: List<T> = this.values,
-        shortName: Char? = this.shortName,
-    ): Option<T> =
-        copy(
-            parameter = parameter.copy(
-                type = type,
-                name = name,
-                description = description,
-                regex = regex,
-                optional = optional,
-                multiple = multiple,
-                values = values,
-            ),
-            shortName = shortName
-        )
-
-    fun keys(): List<String> =
-        listOfNotNull(shortName?.toString(), name)
-
-    override fun summary(): String =
-        format(
-            aliases()
-                .map { if (hasValue) "$it ${parameter.typeName}" else it }
-                .first()
-        )
-
-    override fun definition(): String =
-        aliases().joinToString(", ").let { if (hasValue) "$it ${parameter.typeName}" else it }
-
-    private fun aliases() = listOfNotNull(
-        shortName?.let { "-$it" },
-        name?.let { "--$it" }
-    )
-
-    override fun detail(): String =
-        listOfNotNull(
-            description?.let { if (it.endsWith('.')) it else "$it." },
-            (regex?.pattern ?: parameter.typeName)?.let(::format),
-            values.ifEmpty { null }?.map(Any::toString)?.let { "Default: " + if (multiple) values else value },
-        )
-        .joinToString(" ")
 }
