+++ args/src/main/kotlin/com/hexagonkt/args/Property.kt
@@ -4,23 +4,10 @@ import kotlin.reflect.KClass
 
 sealed interface Property<T : Any> {
     val type: KClass<T>
-    val name: String?
+    val names: LinkedHashSet<String>
     val description: String?
     val regex: Regex?
     val optional: Boolean
     val multiple: Boolean
     val values: List<T>
-    val value: T?
-
-    fun summary(): String
-    fun definition(): String
-    fun detail(): String
-
-    fun format(text: String): String =
-        when {
-            optional && multiple -> "[$text]..."
-            optional -> "[$text]"
-            multiple -> "$text..."
-            else -> text
-        }
 }
