+++ args/src/main/kotlin/com/hexagonkt/args/formatter/ComponentFormatter.kt
@@ -1,0 +1,7 @@
+package com.hexagonkt.args.formatter
+
+interface ComponentFormatter<T> {
+    fun summary(component: T): String
+    fun definition(component: T): String
+    fun detail(component: T): String
+}
